rename app2 to app
