+++ app/adaptive_threshold.py
