+++ app/shot_detector.py
