--- requirements.txt~
@@ -1,4 +1,0 @@
-opencv
-Flask
-werkzeug
-numpy
