+++ run.py
@@ -1,4 +1,4 @@
-from app2.shot_detector import ShotBoundaryDetector
+from app.shot_detector import ShotBoundaryDetector
 import config
 
 import os
